Rename health check package and function names
Rename for consistency
