+++ test/integration/health_test.go
@@ -27,7 +27,7 @@ import (
 	"connectrpc.com/grpchealth"
 	"github.com/stretchr/testify/assert"
 	"github.com/yorkie-team/yorkie/api/yorkie/v1/v1connect"
-	"github.com/yorkie-team/yorkie/server/rpc/health"
+	"github.com/yorkie-team/yorkie/server/rpc/httphealth"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	healthpb "google.golang.org/grpc/health/grpc_health_v1"
@@ -88,7 +88,7 @@ func TestHTTPHealthCheck(t *testing.T) {
 		}()
 		assert.Equal(t, resp.StatusCode, http.StatusOK)
 
-		var healthResp health.CheckResponse
+		var healthResp httphealth.CheckResponse
 		err = json.NewDecoder(resp.Body).Decode(&healthResp)
 		assert.NoError(t, err)
 		assert.Equal(t, healthResp.Status, grpchealth.StatusServing.String())
@@ -106,7 +106,7 @@ func TestHTTPHealthCheck(t *testing.T) {
 			}()
 			assert.Equal(t, resp.StatusCode, http.StatusOK)
 
-			var healthResp health.CheckResponse
+			var healthResp httphealth.CheckResponse
 			err = json.NewDecoder(resp.Body).Decode(&healthResp)
 			assert.NoError(t, err)
 			assert.Equal(t, healthResp.Status, grpchealth.StatusServing.String())
